jdb: enhance {peek,poke}_task() to avoid crossing page boundaries
Both functions use Jdb::access_mem_task() delivering the virtual address
of a mapping. This mapping is usually a single page without a guarantee
that the following mapping points to the following physical page. Hence
we must not cross the page boundary when accessing memory at this
address.

Also change the size type to size_t.

+++ src/lib/disasm/disasm.h
@@ -4,7 +4,7 @@
 #include "globalconfig.h"
 #include "jdb_types.h"
 
-typedef int (*Peek_task)(Jdb_address addr, void *value, int width);
+typedef int (*Peek_task)(Jdb_address addr, void *value, size_t width);
 typedef int (*Is_adp_mem)(Jdb_address addr);
 
 /**
